Utils added, refactoring and renaming of classes
